requirements.txt
Django==4.2.7
python-decouple==3.8
google-generativeai==0.3.2
requests==2.31.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7 